pyproject.toml
[project]
name = "wedding-playlist"
version = "0.1.0"
description = "AI-powered wedding playlist generator using Spotify data and DeepSeek API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "spotipy>=2.23.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]

[project.scripts]
wedding-playlist = "wedding_playlist.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
